requirements.txt
azure-search-documents
azure-identity
langchain-community
langchain-openai
langchain-ollama
python-dotenv
asyncio
streamlit
watchdog
langchain
ollama
